// templateCopy/GoodSleepEIP/ClientApp/src/service/apiServices.schemas.ts
export type JsonResponseData = unknown | null;

export interface JsonResponse {
  success?: boolean;
  /** @nullable */
  message?: string | null;
  /** @nullable */
  data?: JsonResponseData;
  /** @nullable */
  timestamp?: string | null;
}